compiler: support runtime calls in compiler tests
This is a change that I've copied from the build-packages-independently
branch. It makes sure not to assume the parameters slice of functions in
other packages are not already built, instead just relying on the
*types.Signature object.

This is a relatively small change, but it makes it possible to emit
runtime calls in compiler tests, which in turn makes it possible to test
types defined in the runtime and runtime calls (which are often the more
interesting tests).